requirement.txt
boto3==1.9.86
PyYaml==3.13